# [3511] Eager load associations on subjects page