--- a/app/controllers/api/v3/subject_areas_controller.rb
+++ b/app/controllers/api/v3/subject_areas_controller.rb
@@ -2,7 +2,7 @@ module API
   module V3
     class SubjectAreasController < API::V3::ApplicationController
       def index
-        render jsonapi: SubjectArea.active, fields: fields_param, include: params[:include], class: CourseSerializersService.new.execute[:v3]
+        render jsonapi: SubjectArea.active.includes(subjects: [:financial_incentive]), fields: fields_param, include: params[:include], class: CourseSerializersService.new.execute[:v3]
       end
     end
   end