### Context

- https://trello.com/c/6SqF8p2F/3511-optimize-find-courses-by-subject-page
- The subject pages on the find course wizard is not eager loading associations when it hits the api

Before (API)
```
2020-06-09 11:15:24.785430 D [8674:puma threadpool 001] Rails -- API HIT => http://localhost:3001/api/v3/subject_areas?include=subjects
2020-06-09 11:15:24.789229 D [8674:puma threadpool 001] Rack -- Started -- { :method => "GET", :path => "/api/v3/subject_areas?include=subjects", :ip => "::1" }
2020-06-09 11:15:25.048886 D [8674:puma threadpool 001] API::V3::SubjectAreasController -- Processing #index
2020-06-09 11:15:25.069156 D [8674:puma threadpool 001] (0.449ms) ActiveRecord -- SubjectArea Load -- { :sql => "SELECT \"subject_area\".* FROM \"subject_area\" WHERE \"subject_area\".\"typename\" != $1", :binds => { :typename => "DiscontinuedSubject" } }
2020-06-09 11:15:25.084449 D [8674:puma threadpool 001] (0.786ms) ActiveRecord -- Subject Load -- { :sql => "SELECT \"subject\".* FROM \"subject\" WHERE \"subject\".\"type\" = $1", :binds => { :type => "PrimarySubject" } }
2020-06-09 11:15:25.100208 D [8674:puma threadpool 001] (1.015ms) ActiveRecord -- Subject Load -- { :sql => "SELECT \"subject\".* FROM \"subject\" WHERE \"subject\".\"type\" = $1", :binds => { :type => "SecondarySubject" } }
2020-06-09 11:15:25.112032 D [8674:puma threadpool 001] (1.575ms) ActiveRecord -- Subject Load -- { :sql => "SELECT \"subject\".* FROM \"subject\" WHERE \"subject\".\"type\" = $1", :binds => { :type => "ModernLanguagesSubject" } }
2020-06-09 11:15:25.128707 D [8674:puma threadpool 001] (3.293ms) ActiveRecord -- Subject Load -- { :sql => "SELECT \"subject\".* FROM \"subject\" WHERE \"subject\".\"type\" = $1", :binds => { :type => "FurtherEducationSubject" } }
2020-06-09 11:15:25.145438 D [8674:puma threadpool 001] (0.603ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 1, :limit => 1 } }
2020-06-09 11:15:25.147567 D [8674:puma threadpool 001] (1.182ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 2, :limit => 1 } }
2020-06-09 11:15:25.149453 D [8674:puma threadpool 001] (1.161ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 3, :limit => 1 } }
2020-06-09 11:15:25.151702 D [8674:puma threadpool 001] (1.410ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 4, :limit => 1 } }
2020-06-09 11:15:25.171174 D [8674:puma threadpool 001] (1.091ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 5, :limit => 1 } }
2020-06-09 11:15:25.172994 D [8674:puma threadpool 001] (1.012ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 6, :limit => 1 } }
2020-06-09 11:15:25.174776 D [8674:puma threadpool 001] (1.088ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 7, :limit => 1 } }
2020-06-09 11:15:25.176475 D [8674:puma threadpool 001] (0.871ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 8, :limit => 1 } }
2020-06-09 11:15:25.178607 D [8674:puma threadpool 001] (1.150ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 9, :limit => 1 } }
2020-06-09 11:15:25.181304 D [8674:puma threadpool 001] (1.120ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 10, :limit => 1 } }
2020-06-09 11:15:25.183232 D [8674:puma threadpool 001] (0.528ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 11, :limit => 1 } }
2020-06-09 11:15:25.184967 D [8674:puma threadpool 001] (0.689ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 12, :limit => 1 } }
2020-06-09 11:15:25.191831 D [8674:puma threadpool 001] (5.571ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 13, :limit => 1 } }
2020-06-09 11:15:25.196290 D [8674:puma threadpool 001] (3.129ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 14, :limit => 1 } }
2020-06-09 11:15:25.204840 D [8674:puma threadpool 001] (7.171ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 15, :limit => 1 } }
2020-06-09 11:15:25.207166 D [8674:puma threadpool 001] (1.171ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 16, :limit => 1 } }
2020-06-09 11:15:25.208955 D [8674:puma threadpool 001] (0.806ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 17, :limit => 1 } }
2020-06-09 11:15:25.211886 D [8674:puma threadpool 001] (1.855ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 18, :limit => 1 } }
2020-06-09 11:15:25.213597 D [8674:puma threadpool 001] (0.871ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 19, :limit => 1 } }
2020-06-09 11:15:25.215221 D [8674:puma threadpool 001] (0.926ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 20, :limit => 1 } }
2020-06-09 11:15:25.216654 D [8674:puma threadpool 001] (0.603ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 21, :limit => 1 } }
2020-06-09 11:15:25.218306 D [8674:puma threadpool 001] (0.668ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 22, :limit => 1 } }
2020-06-09 11:15:25.219996 D [8674:puma threadpool 001] (0.672ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 23, :limit => 1 } }
2020-06-09 11:15:25.221818 D [8674:puma threadpool 001] (0.973ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 24, :limit => 1 } }
2020-06-09 11:15:25.232255 D [8674:puma threadpool 001] (8.687ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 25, :limit => 1 } }
2020-06-09 11:15:25.234919 D [8674:puma threadpool 001] (0.504ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 26, :limit => 1 } }
2020-06-09 11:15:25.237665 D [8674:puma threadpool 001] (1.654ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 27, :limit => 1 } }
2020-06-09 11:15:25.240439 D [8674:puma threadpool 001] (1.781ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 28, :limit => 1 } }
2020-06-09 11:15:25.242144 D [8674:puma threadpool 001] (0.723ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 29, :limit => 1 } }
2020-06-09 11:15:25.244514 D [8674:puma threadpool 001] (0.772ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 30, :limit => 1 } }
2020-06-09 11:15:25.246552 D [8674:puma threadpool 001] (1.257ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 31, :limit => 1 } }
2020-06-09 11:15:25.249147 D [8674:puma threadpool 001] (1.132ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 32, :limit => 1 } }
2020-06-09 11:15:25.251468 D [8674:puma threadpool 001] (1.551ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 33, :limit => 1 } }
2020-06-09 11:15:25.253314 D [8674:puma threadpool 001] (0.816ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 34, :limit => 1 } }
2020-06-09 11:15:25.261899 D [8674:puma threadpool 001] (7.122ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 35, :limit => 1 } }
2020-06-09 11:15:25.263415 D [8674:puma threadpool 001] (0.500ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 36, :limit => 1 } }
2020-06-09 11:15:25.264778 D [8674:puma threadpool 001] (0.471ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 37, :limit => 1 } }
2020-06-09 11:15:25.266717 D [8674:puma threadpool 001] (1.124ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 38, :limit => 1 } }
2020-06-09 11:15:25.272748 D [8674:puma threadpool 001] (4.158ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 39, :limit => 1 } }
2020-06-09 11:15:25.277757 D [8674:puma threadpool 001] (3.646ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 40, :limit => 1 } }
2020-06-09 11:15:25.282059 D [8674:puma threadpool 001] (2.405ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 41, :limit => 1 } }
2020-06-09 11:15:25.285314 D [8674:puma threadpool 001] (1.881ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 42, :limit => 1 } }
2020-06-09 11:15:25.290432 D [8674:puma threadpool 001] (2.301ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" = $1 LIMIT $2", :binds => { :subject_id => 43, :limit => 1 } }
2020-06-09 11:15:25.384051 I [8674:puma threadpool 001] Rails -- Completed JSON API rendering (316.76ms)
2020-06-09 11:15:25.384287 I [8674:puma threadpool 001] (335.3ms) API::V3::SubjectAreasController -- Completed #index -- { :controller => "API::V3::SubjectAreasController", :action => "index", :params => { "include" => "subjects" }, :format => "JSONAPI", :method => "GET", :path => "/api/v3/subject_areas", :status => 200, :view_runtime => 230.74, :sk_rendered_format => :jsonapi, :sk_variant => [], :db_runtime => 86.29, :status_message => "OK" }
```

Before (Find)
```
2020-06-09 10:53:21.408527 D [40024:puma threadpool 003] Rack -- Started -- {:method=>"GET", :path=>"/start/subject?l=2", :ip=>"::1"}
2020-06-09 10:53:21.410335 D [40024:puma threadpool 003] ResultFilters::SubjectController -- Processing #start
2020-06-09 10:53:21.826747 D [40024:puma threadpool 003] ActionView -- Rendering -- {:template=>"result_filters/subject/start.html.erb"}
2020-06-09 10:53:21.827221 D [40024:puma threadpool 003] (0.048ms) ActionView -- Rendered -- {:partial=>"shared/_error_message.html.erb"}
2020-06-09 10:53:21.828227 D [40024:puma threadpool 003] (0.791ms) ActionView -- Rendered -- {:partial=>"shared/_hidden_fields.html.erb"}
2020-06-09 10:53:21.883158 D [40024:puma threadpool 003] (56.0ms) ActionView -- Rendered -- {:partial=>"result_filters/subject/_form.html.erb"}
2020-06-09 10:53:21.883216 D [40024:puma threadpool 003] (56.4ms) ActionView -- Rendered -- {:template=>"result_filters/subject/start.html.erb", :within=>"layouts/application"}
2020-06-09 10:53:21.890372 I [40024:puma threadpool 003] [Webpacker] Rails -- Everything's up-to-date. Nothing to do
2020-06-09 10:53:21.891476 D [40024:puma threadpool 003] (0.145ms) ActionView -- Rendered -- {:partial=>"layouts/_cookie-banner.html.erb"}
2020-06-09 10:53:21.898380 I [40024:puma threadpool 003] [Webpacker] Rails -- Everything's up-to-date. Nothing to do
2020-06-09 10:53:21.899145 I [40024:puma threadpool 003] (488.7ms) ResultFilters::SubjectController -- Completed #start -- {:controller=>"ResultFilters::SubjectController", :action=>"start", :params=>{"l"=>"2"}, :format=>"HTML", :method=>"GET", :path=>"/start/subject", :status=>200, :view_runtime=>72.6, :request_id=>"6efe5868-a2d3-4689-9521-b183dbc7dce0", :status_message=>"OK"}
```

After (API)
```
2020-06-09 11:10:04.191840 D [8674:puma threadpool 005] Rails -- API HIT => http://localhost:3001/api/v3/subject_areas?include=subjects
2020-06-09 11:10:04.206727 D [8674:puma threadpool 005] Rack -- Started -- { :method => "GET", :path => "/api/v3/subject_areas?include=subjects", :ip => "::1" }
2020-06-09 11:10:04.883354 D [8674:puma threadpool 005] API::V3::SubjectAreasController -- Processing #index
2020-06-09 11:10:04.916829 D [8674:puma threadpool 005] (1.207ms) ActiveRecord -- SubjectArea Load -- { :sql => "SELECT \"subject_area\".* FROM \"subject_area\" WHERE \"subject_area\".\"typename\" != $1", :binds => { :typename => "DiscontinuedSubject" } }
2020-06-09 11:10:04.945207 D [8674:puma threadpool 005] (2.049ms) ActiveRecord -- Subject Load -- { :sql => "SELECT \"subject\".* FROM \"subject\" WHERE \"subject\".\"type\" IN ($1, $2, $3, $4)", :binds => { :type => [ "PrimarySubject", "SecondarySubject", "ModernLanguagesSubject", "FurtherEducationSubject" ] } }
2020-06-09 11:10:05.002940 D [8674:puma threadpool 005] (1.854ms) ActiveRecord -- FinancialIncentive Load -- { :sql => "SELECT \"financial_incentive\".* FROM \"financial_incentive\" WHERE \"financial_incentive\".\"subject_id\" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22, $23, $24, $25, $26, $27, $28, $29, $30, $31, $32, $33, $34, $35, $36, $37, $38, $39, $40, $41, $42, $43)", :binds => { :subject_id => [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 34, 35, 36, 37, 38, 39, 40, 41, 43, 33, 42 ] } }
2020-06-09 11:10:05.031984 I [8674:puma threadpool 005] Rails -- Completed JSON API rendering (120.5ms)
2020-06-09 11:10:05.032323 I [8674:puma threadpool 005] (148.8ms) API::V3::SubjectAreasController -- Completed #index -- { :controller => "API::V3::SubjectAreasController", :action => "index", :params => { "include" => "subjects" }, :format => "JSONAPI", :method => "GET", :path => "/api/v3/subject_areas", :status => 200, :view_runtime => 115.63, :sk_rendered_format => :jsonapi, :sk_variant => [], :db_runtime => 6.05, :status_message => "OK" }
```

After (Find)
```
2020-06-09 10:52:57.364120 D [40024:puma threadpool 005] Rack -- Started -- {:method=>"GET", :path=>"/start/subject?l=2", :ip=>"::1"}
2020-06-09 10:52:57.366056 D [40024:puma threadpool 005] ResultFilters::SubjectController -- Processing #start
2020-06-09 10:52:57.437265 D [40024:puma threadpool 005] ActionView -- Rendering -- {:template=>"result_filters/subject/start.html.erb"}
2020-06-09 10:52:57.437793 D [40024:puma threadpool 005] (0.077ms) ActionView -- Rendered -- {:partial=>"shared/_error_message.html.erb"}
2020-06-09 10:52:57.438743 D [40024:puma threadpool 005] (0.806ms) ActionView -- Rendered -- {:partial=>"shared/_hidden_fields.html.erb"}
2020-06-09 10:52:57.544179 D [40024:puma threadpool 005] (106.5ms) ActionView -- Rendered -- {:partial=>"result_filters/subject/_form.html.erb"}
2020-06-09 10:52:57.544234 D [40024:puma threadpool 005] (106.9ms) ActionView -- Rendered -- {:template=>"result_filters/subject/start.html.erb", :within=>"layouts/application"}
2020-06-09 10:52:57.551611 I [40024:puma threadpool 005] [Webpacker] Rails -- Everything's up-to-date. Nothing to do
…