@asmega 2020-06-09
![](https://github.trello.services/images/mini-trello-icon.png) [[S] optimize Find courses by subject page](https://trello.com/c/6SqF8p2F/3511-optimize-find-courses-by-subject-page)